requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.56.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
threadpoolctl==3.6.0
tld==0.13
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
whois-api==1.2.0
Flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
joblib==1.0.1
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
matplotlib==3.4.2
seaborn==0.11.1
tld==0.12.6
whois==0.9.13
